Rename and relocate some structs in OpenGL.
The C++ module is now named [...].OpenGL.Native.Cpp and the Rust module
is [...].OpenGL.Native.Rust. Also, some missing typespecs were added.
