+++ c_src/opengl_video_compositor/opengl_video_compositor.spec.exs
@@ -1,4 +1,4 @@
-module Membrane.VideoCompositor.OpenGL.Native
+module Membrane.VideoCompositor.OpenGL.Native.Cpp
 
 state_type "State"
