+++ native/membrane_videocompositor_opengl_rust/src/lib.rs
@@ -38,7 +38,7 @@ mod atoms {
 }
 
 #[derive(Debug, rustler::NifStruct)]
-#[module = "Membrane.VideoCompositor.OpenGL.Rust.RawVideo"]
+#[module = "Membrane.VideoCompositor.OpenGL.Native.Rust.RawVideo"]
 struct RawVideo {
     width: usize,
     height: usize,
@@ -287,7 +287,7 @@ impl<T> ResultExt<T> for Result<T, egl::Error> {
 }
 
 rustler::init!(
-    "Elixir.Membrane.VideoCompositor.OpenGL.Rust",
+    "Elixir.Membrane.VideoCompositor.OpenGL.Native.Rust",
     [init, join_frames],
     load = load
 );
